+++ src/cast/message/receiver.rs
@@ -5,7 +5,7 @@ use super::super::proto::{CastMessage, CastMessage_PayloadType, CastMessage_Prot
 
 pub const NAMESPACE: &str = "urn:x-cast:com.google.cast.receiver";
 
-pub fn launch(request_id: i32, app_id: &str) -> Result<CastMessage, Error> {
+pub fn launch(request_id: i64, app_id: &str) -> Result<CastMessage, Error> {
     let payload = to_string(&receiver::Payload::Launch {
         request_id,
         app_id: app_id.to_owned(),
@@ -13,7 +13,7 @@ pub fn launch(request_id: i32, app_id: &str) -> Result<CastMessage, Error> {
     Ok(message(super::DEFAULT_DESTINATION_ID, payload))
 }
 
-pub fn status(request_id: i32) -> Result<CastMessage, Error> {
+pub fn status(request_id: i64) -> Result<CastMessage, Error> {
     let payload = serde_json::to_string(&receiver::Payload::GetStatus { request_id })?;
     Ok(message(super::DEFAULT_DESTINATION_ID, payload))
 }
